Performance Upgrade
Scudo now obtains the return address of the encrypted function and places a temporary breakpoint which will allow the program when to re-encrypts the function immediately after the decrypted function finishes execution.

After doing some performance evaluations using the chrono library, the difference between the old method and new method of re-encrypting the function is about 1060 milliseconds. Meaning, were saving more than a second with this new method over the old.

In terms of performance, calling the function while its encrypted is only 2% slower than when its not.